--- a/sanitize/sanitize_change_test.go
+++ b/sanitize/sanitize_change_test.go
@@ -20,6 +20,11 @@ type testChangeCase struct {
 func changeCases() []testChangeCase {
 	return []testChangeCase{
 		{
+			name:     "nil",
+			old:      nil,
+			expected: nil,
+		},
+		{
 			name: "basic",
 			old: &tfjson.Change{
 				Before: map[string]interface{}{