--- a/sanitize/sanitize_plan_variables.go
+++ b/sanitize/sanitize_plan_variables.go
@@ -26,7 +26,11 @@ func sanitizeVariable(
 	config *tfjson.ConfigVariable,
 	replaceWith interface{},
 ) {
-	if config != nil && config.Sensitive {
+	if result == nil || config == nil {
+		return
+	}
+
+	if config.Sensitive {
 		result.Value = replaceWith
 	}
 }