This PR introduces a breaking change as it changes the library from copying the plan and it's values before sanitizing to overriding the plan with sanitized values. 
The reason for this change is that it improve the performance see the report below.

Report (from `benchstat small_old.txt small_new.txt`)
``` 
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
                  │ small_old.txt │            small_new.txt             │
                  │    sec/op     │    sec/op     vs base                │
LargeChangeset-16   2360.93µ ± 5%   23.92µ ± 10%  -98.99% (p=0.000 n=10)

                  │ small_old.txt  │            small_new.txt             │
                  │      B/op      │     B/op      vs base                │
LargeChangeset-16   895.544Ki ± 0%   7.492Ki ± 0%  -99.16% (p=0.000 n=10)

                  │ small_old.txt │           small_new.txt            │
                  │   allocs/op   │ allocs/op   vs base                │
LargeChangeset-16    29573.0 ± 0%   368.0 ± 0%  -98.76% (p=0.000 n=10)
```

Before this PR (`go test -v ./sanitize/ -bench . -benchmem -run '^$' -count=10 > small_old.txt`)
```
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
BenchmarkLargeChangeset
BenchmarkLargeChangeset-16    	     506	   2467080 ns/op	  917107 B/op	   29574 allocs/op
BenchmarkLargeChangeset-16    	     442	   2483603 ns/op	  916968 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     573	   2331983 ns/op	  917001 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     516	   2309136 ns/op	  917041 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     613	   2347617 ns/op	  916977 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     457	   2374249 ns/op	  917070 B/op	   29574 allocs/op
BenchmarkLargeChangeset-16    	     558	   2197829 ns/op	  917033 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     493	   2438643 ns/op	  917045 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     483	   2482222 ns/op	  917000 B/op	   29573 allocs/op
BenchmarkLargeChangeset-16    	     459	   2297840 ns/op	  917107 B/op	   29574 allocs/op
PASS
ok  	github.com/terramate-io/tfjson/sanitize	15.578s
```

With this PR (`go test -v ./sanitize/ -bench . -benchmem -run '^$' -count=10 > small_new.txt`)
```
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
BenchmarkLargeChangeset
BenchmarkLargeChangeset-16    	   48540	     26195 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   45643	     23769 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   51070	     26817 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   47794	     24536 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   50151	     23856 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   52357	     23988 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   47992	     25676 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   54344	     22974 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   54070	     23603 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   46740	     23639 ns/op	    7672 B/op	     368 allocs/op
PASS
ok  	github.com/terramate-io/tfjson/sanitize	14.742s
```