@boekkooi-terramate 2024-08-23
@i4ki I disagree with your opinion as if you accept a pointer into a (public) func you must accept that users can pass in `nil` as it's a valid value. 
Causing a `panic: runtime error: invalid memory address or nil pointer dereference` is in my opinion a terrible experience.