@boekkooi-terramate 2024-08-23
@i4ki this is also inline with the go standard library