@i4ki 2024-08-23
In C, C++, Go, etc, if a function expects a pointer, it's assumed to be a valid pointer **unless documented that it's an optional argument**.