@i4ki 2024-08-23
If someone does:
```
plan, err = doPlan()
// forget err check
sanitize.SanitizePlanWithValue(plan, err)
```
the panic is better than returning `nil` because it's a programming error.