@boekkooi-terramate 2024-08-23
@i4ki as a web developer you learn that you never trust user input (related to [OWASP Top 10 - Injection](https://owasp.org/Top10/).
Now as the plan is loaded from user input any of it's value's maybe wrong or `nil`. 

The wrong values is partially covered by the unmarshal processes. However, since pointers are used, checking to see if they are `nil` before use it just best practice to avoid any a *panic*.

Also regarding your statement "unless documented that it's an optional argument." I have checked [Effective Go](https://go.dev/doc/effective_go) but can't find any reference to your statement. Could you please provide documentation regarding the statement?

> If someone does:
> 
> ```
> plan, err = doPlan()
> // forget err check
> sanitize.SanitizePlanWithValue(plan, value)
> ```
> 
> the panic is better than returning `nil` because it's a programming error.

Ignoring the error from `doPlan` is a developer mistake. However, if they call `SanitizePlanWithValue` it will return an `error` at the moment instead of causing a `panic` which to me seem to align with "The usual way to report an error to a caller is to return an error as an extra return value." from [Effective Go - Errors - panic](https://go.dev/doc/effective_go#panic).